requirements.txt
# ========================================
# Core ML/AI Libraries
# ========================================
torch==2.5.1, 
torchvision,
torchaudio,
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
datasets>=2.14.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# ========================================
# Database Connectors
# ========================================
psycopg2-binary>=2.9.9  # PostgreSQL
PyMySQL>=1.1.0          # MySQL
pyodbc>=5.0.0           # SQL Server
cx-Oracle>=8.3.0        # Oracle
sqlalchemy>=2.0.23      # Universal ORM

# ========================================
# Data Processing
# ========================================
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlparse>=0.4.4

# ========================================
# API & Web
# ========================================
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ========================================
# Utilities
# ========================================
python-dotenv>=1.0.0
tqdm>=4.66.0
colorlog>=6.8.0
tenacity>=8.2.3

# ========================================
# Development & Testing
# ========================================
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0

# ========================================
# Monitoring
# ========================================
tensorboard>=2.14.0
wandb>=0.16.0

# ========================================
# Document Processing (for documentation)
# ========================================
python-docx>=1.1.0
